requirements.txt
google-cloud-aiplatform
gradio
pdfminer
nltk
langchain
PyPDF2
pdfminer.six